Add unit-tests workflow for pull_request_trigger back
This is simplification for users with write access to have tests started even
without the comment so it will work as expected.

These tests will be no-op for external contributors. For them only first job is
executed and this job only getting information but not testing anything.

This workflow file is taken from master branch where it runs on comment. We need
to copy it here because pull_request_target trigger is triggered from the PR
target branch only.

Related: rhbz#1885635